web_monitor/requirements.txt
flask
pymongo
redis
requests
untangle
zeroconf
pyyaml
opentimestamps-client
